remove async word in subprocess file + rename from sub-process to subprocess + remove unused console.log
